Update Python versions
This is what I am thinking:

Python 3.5 and 3.6 are too old and EOL-ed.  Let us drop them.

Python 3.8 is still supported. Distributions ship Python 3.9
currently, and probably considered the de facto Python for most
purposes.

Python 3.10 is the latest and greatest.  It would be a good idea to
test against it, in order to detect breakage against the future de
facto Python.